@@ -27,41 +27,50 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  intl: ^0.16.0
+  cupertino_icons: ^1.0.2
+  intl: ^0.17.0
+  flutter_image_compress: ^1.0.0
   percent_indicator: "^2.1.6"
-  font_awesome_flutter: ^8.8.1
+  font_awesome_flutter: ^9.0.0
   flutter_rating_bar: ^3.0.1+1
   carousel_slider: ^2.2.1
   qr_flutter: ^3.1.0
   extended_text: ^4.0.0 #TODO REMOVE
   flare_flutter: ^2.0.6
-  pull_to_refresh: 1.6.3
-  optimized_cached_image: ^1.0.0
-  image_picker: ^0.6.7+10
+  pull_to_refresh: 1.6.5
+  optimized_cached_image: ^2.0.1
+  image_picker: ^0.7.4
   inview_notifier_list: ^1.0.0
   page_transition: "^1.1.6"
   equatable: ^1.2.5
-  dio: ^3.0.10
-  bloc: ^6.1.0
-  flutter_bloc: ^6.1.1
-  hydrated_bloc: ^6.1.0
+  dio: ^4.0.0
+  bloc: ^7.0.0
+  flutter_bloc: ^7.0.1
+  hydrated_bloc: ^7.0.1
   auto_animated: any
-  maps_launcher: ^1.2.2+2
-  path_provider: ^1.5.0
+  maps_launcher: ^2.0.0
+  path_provider: ^2.0.1
   share: ^0.6.5+4 #TODO REMOVE
   timeago: '2.0.25'
-  shared_preferences: ^0.5.12+4
-  location: ^3.0.0
+  shared_preferences: ^2.0.6
+  location: ^4.3.0
   flutter_facebook_login: ^3.0.0
   google_sign_in: ^4.5.6
-  firebase_auth: ^0.18.4+1 #TODO REMOVE
-  image_cropper: ^1.3.1
+  # firebase_auth: ^0.18.4+1 #TODO REMOVE
+  image_cropper: ^1.3.0
   flutter_paystack: ^1.0.4+1
-  url_launcher: ^5.7.10
+  url_launcher: ^6.0.6
   random_string: ^2.1.0
-  crypt: ^3.0.1
+  crypt: ^4.0.1
   sentry_flutter: ^4.0.6
+  image_editor_pro: ^1.1.8
+  video_trimmer: ^0.5.1
+  file_picker: ^3.0.2+2
+  video_player: ^2.1.6
+  video_thumbnail: ^0.3.3
+  visibility_detector: ^0.2.0
+  cached_video_player: ^2.0.0-dev.1
+  # firebase_core: ^0.5.3
 
 dev_dependencies:
   flutter_test:
@@ -82,6 +91,7 @@ flutter:
   assets:
    - assets/icons/heart.flr
    - assets/images/main_background.png
+   - assets/images/welcome_image.png
 
 
   # An image asset can refer to one or more resolution-specific "variants", see
