@@ -1,9 +1,12 @@
 import 'dart:convert';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:groovenation_flutter/constants/social_home_type.dart';
 import 'package:groovenation_flutter/constants/user_location_status.dart';
 import 'package:groovenation_flutter/cubit/state/social_state.dart';
 import 'package:groovenation_flutter/data/repo/social_repository.dart';
 import 'package:groovenation_flutter/models/api_result.dart';
+import 'package:groovenation_flutter/models/club.dart';
 import 'package:groovenation_flutter/models/social_person.dart';
 import 'package:groovenation_flutter/models/social_post.dart';
 import 'package:groovenation_flutter/util/location_util.dart';
@@ -15,6 +18,55 @@ class SocialCubit extends HydratedCubit<SocialState> {
 
   final SocialRepository socialRepository;
   final SocialHomeType type;
+
+  void updateSocialPersonIfExists(SocialPerson person) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      List<SocialPost> newList = [];
+
+      socialPosts.forEach((element) {
+        if (element.person.personID == person.personID) {
+          element.person = person;
+        }
+        newList.add(element);
+      });
+
+      emit(SocialLoadedState(
+          socialPosts: socialPosts,
+          hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+    }
+  }
+
+  void updateSocialPostIfExists(SocialPost post) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      int index =
+          socialPosts.indexWhere((element) => element.postID == post.postID);
+
+      if (index != -1) {
+        socialPosts.removeAt(index);
+        socialPosts.insert(index, post);
+
+        print(socialPosts[index].hasUserLiked);
+
+        emit(SocialLoadedState(
+            socialPosts: socialPosts,
+            hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+      }
+    }
+  }
+
+  void insertUserSocialPost(SocialPost post) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      socialPosts.insert(0, post);
+
+      // emit(SocialLoadingState());
+      emit(SocialLoadedState(
+          socialPosts: socialPosts,
+          hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+    }
+  }
 
   void getSocialPosts(int page) async {
     List<SocialPost> socialPosts = [];
@@ -48,7 +100,8 @@ class SocialCubit extends HydratedCubit<SocialState> {
           result = await socialRepository.getTrendingSocialPosts(page);
           break;
         case SocialHomeType.USER:
-          result = await socialRepository.getProfileSocialPosts(page, sharedPrefs.userId);
+          result = await socialRepository.getProfileSocialPosts(
+              page, sharedPrefs.userId);
           break;
         default:
       }
@@ -114,6 +167,105 @@ class ProfileSocialCubit extends Cubit<SocialState> {
 
   final SocialRepository socialRepository;
 
+  void updateSocialPersonIfExists(SocialPerson person) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      List<SocialPost> newList = [];
+
+      socialPosts.forEach((element) {
+        if (element.person.personID == person.personID) {
+          element.person = person;
+        }
+        newList.add(element);
+      });
+
+      emit(SocialLoadedState(
+          socialPosts: socialPosts,
+          hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+    }
+  }
+
+  void updateUserFollowing(BuildContext context, SocialPerson person) async {
+
+    try {
+      final FollowingSocialCubit followingSocialCubit =
+          BlocProvider.of<FollowingSocialCubit>(context);
+      final TrendingSocialCubit trendingSocialCubit =
+          BlocProvider.of<TrendingSocialCubit>(context);
+      final NearbySocialCubit nearbySocialCubit =
+          BlocProvider.of<NearbySocialCubit>(context);
+      final UserSocialCubit userSocialCubit =
+          BlocProvider.of<UserSocialCubit>(context);
+      final ProfileSocialCubit profileSocialCubit =
+          BlocProvider.of<ProfileSocialCubit>(context);
+      final ClubMomentsCubit clubMomentsCubit =
+          BlocProvider.of<ClubMomentsCubit>(context);
+
+      followingSocialCubit.updateSocialPersonIfExists(person);
+      trendingSocialCubit.updateSocialPersonIfExists(person);
+      nearbySocialCubit.updateSocialPersonIfExists(person);
+      userSocialCubit.updateSocialPersonIfExists(person);
+      profileSocialCubit.updateSocialPersonIfExists(person);
+      clubMomentsCubit.updateSocialPersonIfExists(person);
+
+      await socialRepository.changeUserFollowing(person);
+
+      emit(SocialPostLikeSuccessState());
+    } on SocialException catch (e) {
+      person.isUserFollowing = !person.isUserFollowing;
+
+      final FollowingSocialCubit followingSocialCubit =
+          BlocProvider.of<FollowingSocialCubit>(context);
+      final TrendingSocialCubit trendingSocialCubit =
+          BlocProvider.of<TrendingSocialCubit>(context);
+      final NearbySocialCubit nearbySocialCubit =
+          BlocProvider.of<NearbySocialCubit>(context);
+      final UserSocialCubit userSocialCubit =
+          BlocProvider.of<UserSocialCubit>(context);
+      final ProfileSocialCubit profileSocialCubit =
+          BlocProvider.of<ProfileSocialCubit>(context);
+      final ClubMomentsCubit clubMomentsCubit =
+          BlocProvider.of<ClubMomentsCubit>(context);
+
+      followingSocialCubit.updateSocialPersonIfExists(person);
+      trendingSocialCubit.updateSocialPersonIfExists(person);
+      nearbySocialCubit.updateSocialPersonIfExists(person);
+      userSocialCubit.updateSocialPersonIfExists(person);
+      profileSocialCubit.updateSocialPersonIfExists(person);
+      clubMomentsCubit.updateSocialPersonIfExists(person);
+
+      emit(SocialPostLikeErrorState(e.error));
+    }
+  }
+
+  void updateSocialPostIfExists(SocialPost post) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      int index =
+          socialPosts.indexWhere((element) => element.postID == post.postID);
+
+      if (index != -1) {
+        socialPosts.removeAt(index);
+        socialPosts.insert(index, post);
+
+        emit(SocialLoadedState(
+            socialPosts: socialPosts,
+            hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+      }
+    }
+  }
+
+  void insertUserSocialPost(SocialPost post) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      socialPosts.insert(0, post);
+
+      emit(SocialLoadedState(
+          socialPosts: socialPosts,
+          hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+    }
+  }
+
   void getSocialPosts(int page, SocialPerson socialPerson) async {
     emit(SocialLoadingState());
     try {
@@ -140,6 +292,41 @@ class ClubMomentsCubit extends Cubit<SocialState> {
 
   final SocialRepository socialRepository;
 
+  void updateSocialPersonIfExists(SocialPerson person) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      List<SocialPost> newList = [];
+
+      socialPosts.forEach((element) {
+        if (element.person.personID == person.personID) {
+          element.person = person;
+        }
+        newList.add(element);
+      });
+
+      emit(SocialLoadedState(
+          socialPosts: socialPosts,
+          hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+    }
+  }
+
+  void updateSocialPostIfExists(SocialPost post) {
+    if (state is SocialLoadedState) {
+      List<SocialPost> socialPosts = (state as SocialLoadedState).socialPosts;
+      int index =
+          socialPosts.indexWhere((element) => element.postID == post.postID);
+
+      if (index != -1) {
+        socialPosts.removeAt(index);
+        socialPosts.insert(index, post);
+
+        emit(SocialLoadedState(
+            socialPosts: socialPosts,
+            hasReachedMax: (state as SocialLoadedState).hasReachedMax));
+      }
+    }
+  }
+
   void getMoments(int page, String clubId) async {
     emit(SocialLoadingState());
     try {
@@ -156,6 +343,109 @@ class ClubMomentsCubit extends Cubit<SocialState> {
           socialPosts: newSocial, hasReachedMax: hasReachedMax));
     } on SocialException catch (e) {
       emit(SocialErrorState(e.error));
+    }
+  }
+}
+
+class SocialPostCubit extends Cubit<SocialState> {
+  SocialPostCubit(this.socialRepository) : super(SocialInitialState());
+
+  final SocialRepository socialRepository;
+
+  void uploadPost(BuildContext context, String mediaFilePath, String caption,
+      String clubId, bool isVideo) async {
+    emit(SocialPostUploadLoadingState());
+    try {
+      SocialPost post = await socialRepository.uploadSocialPost(
+          mediaFilePath, caption, clubId, isVideo);
+
+      final FollowingSocialCubit followingSocialCubit =
+          BlocProvider.of<FollowingSocialCubit>(context);
+      final NearbySocialCubit nearbySocialCubit =
+          BlocProvider.of<NearbySocialCubit>(context);
+      final UserSocialCubit userSocialCubit =
+          BlocProvider.of<UserSocialCubit>(context);
+
+      followingSocialCubit.insertUserSocialPost(post);
+      if (post.clubID != null) nearbySocialCubit.insertUserSocialPost(post);
+      userSocialCubit.insertUserSocialPost(post);
+
+      emit(SocialPostUploadSuccessState(post));
+    } on SocialException catch (e) {
+      emit(SocialPostUploadErrorState(e.error));
+    }
+  }
+
+  void changeLikePost(BuildContext context, SocialPost post) async {
+    emit(SocialPostLikeLoadingState());
+
+    if (post.hasUserLiked) {
+      post.hasUserLiked = false;
+      post.likesAmount = post.likesAmount - 1;
+    } else {
+      post.hasUserLiked = true;
+      post.likesAmount = post.likesAmount + 1;
+    }
+
+    try {
+      emit(SocialPostLikeUpdatingState(post));
+      emit(SocialPostLikeLoadingState());
+
+      final FollowingSocialCubit followingSocialCubit =
+          BlocProvider.of<FollowingSocialCubit>(context);
+      final TrendingSocialCubit trendingSocialCubit =
+          BlocProvider.of<TrendingSocialCubit>(context);
+      final NearbySocialCubit nearbySocialCubit =
+          BlocProvider.of<NearbySocialCubit>(context);
+      final UserSocialCubit userSocialCubit =
+          BlocProvider.of<UserSocialCubit>(context);
+      final ProfileSocialCubit profileSocialCubit =
+          BlocProvider.of<ProfileSocialCubit>(context);
+      final ClubMomentsCubit clubMomentsCubit =
+          BlocProvider.of<ClubMomentsCubit>(context);
+
+      followingSocialCubit.updateSocialPostIfExists(post);
+      trendingSocialCubit.updateSocialPostIfExists(post);
+      nearbySocialCubit.updateSocialPostIfExists(post);
+      userSocialCubit.updateSocialPostIfExists(post);
+      profileSocialCubit.updateSocialPostIfExists(post);
+      clubMomentsCubit.updateSocialPostIfExists(post);
+
+      await socialRepository.changeLikeSocialPost(post);
+
+      emit(SocialPostLikeSuccessState());
+    } on SocialException catch (e) {
+      if (post.hasUserLiked) {
+        post.hasUserLiked = false;
+        post.likesAmount = post.likesAmount - 1;
+      } else {
+        post.hasUserLiked = true;
+        post.likesAmount = post.likesAmount + 1;
+      }
+
+      emit(SocialPostLikeUpdatingState(post));
+
+      final FollowingSocialCubit followingSocialCubit =
+          BlocProvider.of<FollowingSocialCubit>(context);
+      final TrendingSocialCubit trendingSocialCubit =
+          BlocProvider.of<TrendingSocialCubit>(context);
+      final NearbySocialCubit nearbySocialCubit =
+          BlocProvider.of<NearbySocialCubit>(context);
+      final UserSocialCubit userSocialCubit =
+          BlocProvider.of<UserSocialCubit>(context);
+      final ProfileSocialCubit profileSocialCubit =
+          BlocProvider.of<ProfileSocialCubit>(context);
+      final ClubMomentsCubit clubMomentsCubit =
+          BlocProvider.of<ClubMomentsCubit>(context);
+
+      followingSocialCubit.updateSocialPostIfExists(post);
+      trendingSocialCubit.updateSocialPostIfExists(post);
+      nearbySocialCubit.updateSocialPostIfExists(post);
+      userSocialCubit.updateSocialPostIfExists(post);
+      profileSocialCubit.updateSocialPostIfExists(post);
+      clubMomentsCubit.updateSocialPostIfExists(post);
+
+      emit(SocialPostLikeErrorState(e.error));
     }
   }
 }
