@@ -1,5 +1,6 @@
 import 'package:crypt/crypt.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/material.dart';
 import 'package:groovenation_flutter/constants/error.dart';
 import 'package:groovenation_flutter/constants/strings.dart';
 import 'package:groovenation_flutter/models/api_result.dart';
@@ -29,6 +30,7 @@ class SocialRepository {
   }
 
   Future<APIResult> getTrendingSocialPosts(int page) async {
+    print(page);
     var uid = sharedPrefs.userId;
     return getSocialPosts("$API_HOST/social/trending/$uid/$page");
   }
@@ -75,9 +77,11 @@ class SocialRepository {
 
     try {
       Response response =
-          await Dio().get("$API_HOST/social/post/comments/$uid/$postID");
+          await Dio().get("$API_HOST/social/post/comments/$uid/$postID/$page");
       if (response.statusCode == 200) {
         Map<String, dynamic> jsonResponse = response.data;
+
+        print(response);
 
         if (jsonResponse['status'] == 1) {
           bool hasReachedMax = jsonResponse['has_reached_max'];
@@ -97,6 +101,104 @@ class SocialRepository {
         throw SocialException(e.error);
       else
         throw SocialException(Error.NETWORK_ERROR);
+    }
+  }
+
+  Future<SocialPost> uploadSocialPost(
+      String mediaFilePath, String caption, String clubId, bool isVideo) async {
+    var uid = sharedPrefs.userId;
+
+    try {
+      FormData formData = new FormData.fromMap({
+        "user_id": uid,
+        "caption": caption,
+        "club_id": clubId,
+        "social_file": await MultipartFile.fromFile(mediaFilePath,
+            filename: randomAlphaNumeric(15) +
+                DateTime.now().millisecondsSinceEpoch.toString() +
+                p.extension(mediaFilePath)),
+      });
+
+      Response response = await Dio().post("$API_HOST/social/post/create",
+          data: formData, options: Options(contentType: 'multipart/form-data'));
+
+      if (response.statusCode == 200) {
+        Map<String, dynamic> jsonResponse = response.data;
+
+        print(response.data.toString());
+
+        if (jsonResponse['status'] == 1) {
+          if (jsonResponse['social_post'] != null) {
+            SocialPost post = SocialPost.fromJson(jsonResponse['social_post']);
+            return post;
+          }
+
+          return null;
+        } else
+          throw SocialException(Error.UNKNOWN_ERROR);
+      } else
+        throw SocialException(Error.UNKNOWN_ERROR);
+    } catch (e) {
+      print(e);
+      if (e is SocialException)
+        throw SocialException(e.error);
+      else
+        throw SocialException(Error.NETWORK_ERROR);
+    }
+  }
+
+  CancelToken _searchCancelToken;
+  Future<APIResult> searchUsers(String searchTerm, int page) async {
+    List<SocialPerson> people = [];
+
+    if (_searchCancelToken != null) {
+      try {
+        _searchCancelToken.cancel();
+        _searchCancelToken = null;
+      } catch (e) {}
+    }
+
+    _searchCancelToken = CancelToken();
+
+    try {
+      Response response = await Dio().post("$API_HOST/social/search/users",
+          data: {
+            'search_term': searchTerm,
+            'page': page,
+            'user_id': sharedPrefs.userId,
+          },
+          cancelToken: _searchCancelToken);
+
+      if (response.statusCode == 200) {
+        Map<String, dynamic> jsonResponse = response.data;
+
+        print(response.data);
+
+        if (jsonResponse['status'] == 1) {
+          bool hasReachedMax = jsonResponse['has_reached_max'];
+
+          for (Map i in jsonResponse['social_people']) {
+            SocialPerson person = SocialPerson.fromJson(i);
+            people.add(person);
+          }
+
+          return APIResult(people, hasReachedMax);
+        } else
+          throw SocialException(Error.UNKNOWN_ERROR);
+      } else
+        throw SocialException(Error.UNKNOWN_ERROR);
+    } catch (e) {
+      print(e);
+      if (e is SocialException)
+        throw SocialException(e.error);
+      else {
+        if (e is DioError) if (e.type == DioErrorType.cancel) {
+          throw e;
+        } else
+          throw SocialException(Error.NETWORK_ERROR);
+        else
+          throw SocialException(Error.NETWORK_ERROR);
+      }
     }
   }
 
@@ -157,11 +259,13 @@ class SocialRepository {
 
     try {
       Response response =
-          await Dio().post("$API_HOST/social/profile/password/change", data: {
-        "userId": userId,
-        "oldPassword": c1.hash,
-        "newPassword": c2.hash,
-      }, options: Options(contentType: Headers.formUrlEncodedContentType));
+          await Dio().post("$API_HOST/social/profile/password/change",
+              data: {
+                "userId": userId,
+                "oldPassword": c1.hash,
+                "newPassword": c2.hash,
+              },
+              options: Options(contentType: Headers.formUrlEncodedContentType));
 
       if (response.statusCode == 200) {
         Map<String, dynamic> jsonResponse = response.data;
@@ -240,7 +344,151 @@ class SocialRepository {
       if (e is SocialException)
         throw SocialException(e.error);
       else {
-        if (e is DioError) if (e.type == DioErrorType.CANCEL) {
+        if (e is DioError) if (e.type == DioErrorType.cancel) {
+          throw e;
+        } else
+          throw SocialException(Error.NETWORK_ERROR);
+        else
+          throw SocialException(Error.NETWORK_ERROR);
+      }
+    }
+  }
+
+  Future<bool> changeLikeSocialPost(SocialPost post) async {
+    var uid = sharedPrefs.userId;
+
+    try {
+      Response response = await Dio().post(
+          post.hasUserLiked
+              ? "$API_HOST/social/post/like"
+              : "$API_HOST/social/post/unlike",
+          data: {
+            "user_id": uid,
+            "post_id": post.postID,
+          });
+      if (response.statusCode == 200) {
+        Map<String, dynamic> jsonResponse = response.data;
+
+        print(response);
+
+        if (jsonResponse['status'] == 1) {
+          return true;
+        } else
+          throw SocialException(Error.UNKNOWN_ERROR);
+      } else
+        throw SocialException(Error.UNKNOWN_ERROR);
+    } catch (e) {
+      if (e is SocialException)
+        throw SocialException(e.error);
+      else {
+        if (e is DioError) if (e.type == DioErrorType.cancel) {
+          throw e;
+        } else
+          throw SocialException(Error.NETWORK_ERROR);
+        else
+          throw SocialException(Error.NETWORK_ERROR);
+      }
+    }
+  }
+  
+  Future<bool> changeUserFollowing(SocialPerson person) async {
+    var uid = sharedPrefs.userId;
+
+    try {
+      Response response = await Dio().get(
+          person.isUserFollowing
+              ? "$API_HOST/social/follow/${person.personID}/$uid"
+              : "$API_HOST/social/unfollow/${person.personID}/$uid");
+              
+      if (response.statusCode == 200) {
+        Map<String, dynamic> jsonResponse = response.data;
+
+        print(response);
+
+        if (jsonResponse['status'] == 1) {
+          return true;
+        } else
+          throw SocialException(Error.UNKNOWN_ERROR);
+      } else
+        throw SocialException(Error.UNKNOWN_ERROR);
+    } catch (e) {
+      if (e is SocialException)
+        throw SocialException(e.error);
+      else {
+        if (e is DioError) if (e.type == DioErrorType.cancel) {
+          throw e;
+        } else
+          throw SocialException(Error.NETWORK_ERROR);
+        else
+          throw SocialException(Error.NETWORK_ERROR);
+      }
+    }
+  }
+
+  Future<bool> changeLikeComment(SocialComment comment) async {
+    var uid = sharedPrefs.userId;
+
+    try {
+      Response response = await Dio().post(
+          comment.hasUserLiked
+              ? "$API_HOST/social/post/comment/like"
+              : "$API_HOST/social/post/comment/unlike",
+          data: {
+            "user_id": uid,
+            "comment_id": comment.commentId,
+          });
+      if (response.statusCode == 200) {
+        Map<String, dynamic> jsonResponse = response.data;
+
+        print(response);
+
+        if (jsonResponse['status'] == 1) {
+          return true;
+        } else
+          throw SocialException(Error.UNKNOWN_ERROR);
+      } else
+        throw SocialException(Error.UNKNOWN_ERROR);
+    } catch (e) {
+      if (e is SocialException)
+        throw SocialException(e.error);
+      else {
+        if (e is DioError) if (e.type == DioErrorType.cancel) {
+          throw e;
+        } else
+          throw SocialException(Error.NETWORK_ERROR);
+        else
+          throw SocialException(Error.NETWORK_ERROR);
+      }
+    }
+  }
+
+  Future<bool> addSocialComment(String postId, String comment) async {
+    var uid = sharedPrefs.userId;
+
+    try {
+      Response response =
+          await Dio().post("$API_HOST/social/post/comment/add", data: {
+        "user_id": uid,
+        "post_id": postId,
+        "comment": comment,
+      });
+
+      if (response.statusCode == 200) {
+        Map<String, dynamic> jsonResponse = response.data;
+
+        print(response);
+
+        if (jsonResponse['status'] == 1) {
+          return true;
+        } else
+          throw SocialException(Error.UNKNOWN_ERROR);
+      } else
+        throw SocialException(Error.UNKNOWN_ERROR);
+    } catch (e) {
+      if (e is SocialException)
+        throw SocialException(e.error);
+      else {
+        if (e is DioError) if (e.type == DioErrorType.cancel) {
           throw e;
         } else
           throw SocialException(Error.NETWORK_ERROR);
