@@ -2,12 +2,12 @@ import 'package:groovenation_flutter/models/social_person.dart';
 
 class SocialPost {
   final String postID; 
-  final SocialPerson person;
+  SocialPerson person;
   final String clubID; //Nullable
   final String clubName; //Nullable
-  final int likesAmount;
+  int likesAmount;
   final String caption;
-  final bool hasUserLiked;
+  bool hasUserLiked;
 
   final int postType;
   final String mediaURL;
