@@ -8,7 +8,7 @@ import 'package:groovenation_flutter/cubit/state/user_cubit_state.dart';
 import 'package:groovenation_flutter/cubit/user_cubit.dart';
 import 'package:groovenation_flutter/models/social_person.dart';
 import 'package:groovenation_flutter/util/alert_util.dart';
-import 'package:optimized_cached_image/widgets.dart';
+import 'package:optimized_cached_image/optimized_cached_image.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class FollowingPage extends StatefulWidget {
