@@ -1,12 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:groovenation_flutter/cubit/social_comments_cubit.dart';
+import 'package:groovenation_flutter/cubit/state/social_comments_state.dart';
 import 'package:groovenation_flutter/models/social_comment.dart';
 import 'package:groovenation_flutter/models/social_person.dart';
+import 'package:groovenation_flutter/util/shared_prefs.dart';
 import 'package:optimized_cached_image/optimized_cached_image.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
 class PostCommentItem extends StatefulWidget {
   final SocialComment socialComment;
-  PostCommentItem(this.socialComment);
+  PostCommentItem(Key key, this.socialComment) : super(key: key);
 
   @override
   _PostCommentItemState createState() => _PostCommentItemState(socialComment);
@@ -14,7 +18,7 @@ class PostCommentItem extends StatefulWidget {
 
 class _PostCommentItemState extends State<PostCommentItem> {
   bool currentMaxLines = false;
-  final SocialComment socialComment;
+  SocialComment socialComment;
   _PostCommentItemState(this.socialComment);
 
   @override
@@ -28,12 +32,26 @@ class _PostCommentItemState extends State<PostCommentItem> {
   }
 
   _likeComment() {
-    //TODO Like Comment
+    if(socialComment.commentId != null){
+      final SocialCommentsLikeCubit socialCommentsLikeCubit =
+          BlocProvider.of<SocialCommentsLikeCubit>(context);
+      socialCommentsLikeCubit.changeLikeComment(context, socialComment);
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return FlatButton(
+    return BlocListener<SocialCommentsLikeCubit, SocialCommentsState>(
+        listener: (context, socialState) {
+          if (socialState is SocialCommentLikeUpdatingState) {
+            if (socialState.comment.commentId == socialComment.commentId){
+              setState(() {
+                socialComment = socialState.comment;
+              });
+            }
+          }
+        },
+        child: FlatButton(
         onPressed: () {
           setState(() {
             currentMaxLines = !currentMaxLines;
@@ -101,7 +119,7 @@ class _PostCommentItemState extends State<PostCommentItem> {
                       ],
                     )),
               ),
-              IconButton(
+              Visibility(visible: socialComment.person.personID != sharedPrefs.userId, child: IconButton(
                   padding: EdgeInsets.zero,
                   icon: Icon(
                     socialComment.hasUserLiked
@@ -112,7 +130,7 @@ class _PostCommentItemState extends State<PostCommentItem> {
                   iconSize: 20,
                   onPressed: () {
                     _likeComment();
-                  })
-            ]));
+                  }))
+            ])));
   }
 }
